Initial Commit
I'm too lazy to split all this up given it's just a spike...

This creates a server that watches a directory on your machine, listens
for changes to anything in that directory and then publishes changes
to that folder on a websocket connection to anything that's connected.

Next step is to have file changes appear on the stream overlay, rather
than print to the console.